task3_image_captioning/requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
transformers>=4.28.0
torch>=2.0.0
